refactor: split up acl tests
