+++ source/dea-app/package.json
@@ -107,6 +107,7 @@
     "ts-mockito": "~2.6.1",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.2",
-    "wait-port": "~1.0.4"
+    "wait-port": "~1.0.4",
+    "ulid": "~2.3.0"
   }
 }
